refactored login tests
